+++ server.ts
@@ -2,7 +2,7 @@ import { createRequestHandler, type ServerBuild } from 'react-router'
 // @ts-ignore This file won’t exist if it hasn’t yet been built
 import * as build from './build/server'
 import { getLoadContext } from './load-context'
-import { queue } from './queue'
+import { queue } from './workers/queue'
 
 const requestHandler = createRequestHandler(build as unknown as ServerBuild)
